fix(infra): address second-round PR review concerns
- Extract shared advisory-lock migration logic into migrate-core.ts;
  migrate-prod.ts and migrate-docker.ts now delegate to it so lock/retry
  behavior stays in sync across deployment paths
- Make POSTGRES_USER/PASSWORD/DB env-var overridable in docker-compose.yml
  (was hardcoded); add comment that DATABASE_URL must be set to match
- Add `image: tokscale:latest` to the compose app service so `make up`
  uses the pre-built image from `make docker/build` and never triggers
  a compose-managed build (which would fail: `db` hostname is only
  resolvable inside the compose network, not the Docker build context)
- Clarify `make docker/build` success message; copy migrate-core.ts in
  the Dockerfile runner stage
